requirements.txt
Appium-Python-Client==0.40
selenium==3.141.0
urllib3==1.24.1
